+++ src/python/webmentions/handlers/_incoming.py
@@ -22,6 +22,7 @@ class IncomingWebmentionsProcessor:  # pylint: disable=too-few-public-methods
         storage: WebmentionsStorage,
         *,
         base_url: str | None = None,
+        base_urls: list[str] | None = None,
         http_timeout: float = DEFAULT_HTTP_TIMEOUT,
         user_agent: str = DEFAULT_USER_AGENT,
         on_mention_processed=None,
@@ -30,7 +31,10 @@ class IncomingWebmentionsProcessor:  # pylint: disable=too-few-public-methods
         **_,
     ):
         self.parser = WebmentionsRequestParser(
-            base_url=base_url, http_timeout=http_timeout, user_agent=user_agent
+            base_url=base_url,
+            base_urls=base_urls,
+            http_timeout=http_timeout,
+            user_agent=user_agent,
         )
         self._storage = storage
         self._on_mention_processed = on_mention_callback_wrapper(on_mention_processed)